Refactor utilities functions to create and read files
