+++ sdks/python/apache_beam/examples/complete/estimate_pi_it_test.py
@@ -27,21 +27,7 @@ import pytest
 
 from apache_beam.examples.complete import estimate_pi
 from apache_beam.testing.test_pipeline import TestPipeline
-
-# Protect against environments where gcsio library is not available.
-try:
-  from apache_beam.io.gcp import gcsio
-except ImportError:
-  gcsio = None
-
-
-def read_gcs_output_file(file_pattern):
-  gcs = gcsio.GcsIO()
-  file_names = gcs.list_prefix(file_pattern).keys()
-  output = []
-  for file_name in file_names:
-    output.append(gcs.open(file_name).read().decode('utf-8'))
-  return '\n'.join(output)
+from apache_beam.testing.test_utils import read_gcs_output_file
 
 
 class EstimatePiIT(unittest.TestCase):
