+++ sdks/python/apache_beam/examples/cookbook/mergecontacts_test.py
@@ -27,29 +27,7 @@ import pytest
 
 from apache_beam.examples.cookbook import mergecontacts
 from apache_beam.testing.test_pipeline import TestPipeline
-
-# Protect against environments where gcsio library is not available.
-try:
-  from apache_beam.io.gcp import gcsio
-except ImportError:
-  gcsio = None
-
-
-def read_gcs_output_file(file_pattern):
-  gcs = gcsio.GcsIO()
-  file_names = gcs.list_prefix(file_pattern).keys()
-  output = []
-  for file_name in file_names:
-    output.append(gcs.open(file_name).read().decode('utf-8').strip())
-  return '\n'.join(output)
-
-
-def create_content_input_file(path, contents):
-  logging.info('Creating file: %s', path)
-  gcs = gcsio.GcsIO()
-  with gcs.open(path, 'w') as f:
-    f.write(str.encode(contents, 'utf-8'))
-  return path
+from apache_beam.testing.test_utils import create_file, read_gcs_output_file
 
 
 class MergeContactsTest(unittest.TestCase):
@@ -163,11 +141,11 @@ class MergeContactsTest(unittest.TestCase):
     # Setup the files with expected content.
     temp_location = test_pipeline.get_option('temp_location')
     input_folder = '/'.join([temp_location, str(uuid.uuid4())])
-    path_email = create_content_input_file(
+    path_email = create_file(
         '/'.join([input_folder, 'path_email.txt']), self.CONTACTS_EMAIL)
-    path_phone = create_content_input_file(
+    path_phone = create_file(
         '/'.join([input_folder, 'path_phone.txt']), self.CONTACTS_PHONE)
-    path_snailmail = create_content_input_file(
+    path_snailmail = create_file(
         '/'.join([input_folder, 'path_snailmail.txt']), self.CONTACTS_SNAILMAIL)
 
     result_prefix = '/'.join([temp_location, str(uuid.uuid4()), 'result'])
