+++ sdks/python/apache_beam/examples/cookbook/multiple_output_pardo_test.py
@@ -28,29 +28,7 @@ import pytest
 
 from apache_beam.examples.cookbook import multiple_output_pardo
 from apache_beam.testing.test_pipeline import TestPipeline
-
-# Protect against environments where gcsio library is not available.
-try:
-  from apache_beam.io.gcp import gcsio
-except ImportError:
-  gcsio = None
-
-
-def read_gcs_output_file(file_pattern):
-  gcs = gcsio.GcsIO()
-  file_names = gcs.list_prefix(file_pattern).keys()
-  output = []
-  for file_name in file_names:
-    output.append(gcs.open(file_name).read().decode('utf-8').strip())
-  return '\n'.join(output)
-
-
-def create_content_input_file(path, contents):
-  logging.info('Creating file: %s', path)
-  gcs = gcsio.GcsIO()
-  with gcs.open(path, 'w') as f:
-    f.write(str.encode(contents, 'utf-8'))
-  return path
+from apache_beam.testing.test_utils import create_file, read_gcs_output_file
 
 
 class MultipleOutputParDo(unittest.TestCase):
@@ -76,8 +54,7 @@ class MultipleOutputParDo(unittest.TestCase):
     # Setup the files with expected content.
     temp_location = test_pipeline.get_option('temp_location')
     input_folder = '/'.join([temp_location, str(uuid.uuid4())])
-    input = create_content_input_file(
-        '/'.join([input_folder, 'input.txt']), self.SAMPLE_TEXT)
+    input = create_file('/'.join([input_folder, 'input.txt']), self.SAMPLE_TEXT)
     result_prefix = '/'.join([temp_location, str(uuid.uuid4()), 'result'])
 
     extra_opts = {'input': input, 'output': result_prefix}
